requirements.txt
streamlit>=1.51.0
PyPDF2==3.0.1
python-dotenv>=1.0.0
groq>=0.36.0
pdfplumber>=0.10.3
